Seperate reducer selectedUnit from unit reducer
