+++ server/server.js
@@ -13,7 +13,7 @@ import rootReducer from '../src/rootReducer';
 import App from '../src/App';
 import {makeLanguageHandler, makeUnitHandler} from './utils';
 import { setLocale } from '../src/redux/actions/locale';
-import { changeSelectedUnit } from '../src/redux/actions/unit';
+import { changeSelectedUnit } from '../src/redux/actions/selectedUnit';
 import { SheetsRegistry } from 'jss';
 import { createGenerateClassName, MuiThemeProvider } from '@material-ui/core';
 import JssProvider from 'react-jss/lib/JssProvider';
