+++ src/redux/selectors/selectedUnit.js
@@ -1,0 +1,3 @@
+export const getSelectedUnit = store => store.selectedUnit.data;
+
+export default getSelectedUnit;
