--- src/redux/selectors/unit.js
@@ -1,1 +1,0 @@
-export const getSelectedUnit = store => store.units.selectedUnit;
