+++ src/views/Map/MapContainer.js
@@ -6,7 +6,7 @@ import { getMapType } from '../../redux/selectors/map';
 import getDistricts from '../../redux/selectors/district';
 import { fetchDistrictsData } from '../../redux/actions/district';
 import MapView from './components/MapView';
-import { getSelectedUnit } from '../../redux/selectors/unit';
+import { getSelectedUnit } from '../../redux/selectors/selectedUnit';
 import { getLocaleString } from '../../redux/selectors/locale';
 import CreateMap from './utils/createMap';
 import { mapOptions } from './constants/mapConstants';
