+++ src/views/SearchView/index.js
@@ -1,14 +1,14 @@
 import { connect } from 'react-redux';
 import SearchView from './SearchView';
 import { fetchUnits } from '../../redux/actions/unit';
-import { changeSelectedUnit } from '../../redux/actions/filter';
+import { changeSelectedUnit } from '../../redux/actions/selectedUnit';
 
 // Listen to redux state
 // const unitList = getUnitList(state);
 const mapStateToProps = (state) => {
   const { units } = state;
   const {
-    data, isFetching, count, max,
+    data, isFetching, count, max, previousSearch,
   } = units;
   return {
     unit: state.unit,
@@ -16,6 +16,7 @@ const mapStateToProps = (state) => {
     isFetching,
     count,
     max,
+    previousSearch,
   };
 };
